Got STEAM loaded!
Got steam set up. Think we are going to have to dl some plugins to handle the steam side of things... we will see.

To see it work, have steam running. Select Stand Alone game and launch
if it works you should see the Shift+Tab to launch Steam in the bot right corner